#12 feat : 스케줄 삭제 후 갱신 기능 추가
