+++ src/pages/schedulePage/components/Timeslot.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { DayObject } from 'types/customTypes';
 import { deleteScheduleAPI } from 'libs/api/schedule';
+import { RefreshedContext } from 'libs/context';
 
 interface TimeslotProps {
   isSpread: boolean;
@@ -11,6 +12,7 @@ interface TimeslotProps {
 export default function Timeslot({ isSpread, day, dayObject }: TimeslotProps) {
   const [needHidden, setNeedHidden] = React.useState<boolean>(false);
   const [isResized, setIsResized] = React.useState<boolean>(false);
+  const { changeRefreshState } = React.useContext(RefreshedContext);
   const { startTime, endTime, id } = dayObject;
 
   React.useEffect(() => {
@@ -51,6 +53,7 @@ export default function Timeslot({ isSpread, day, dayObject }: TimeslotProps) {
       const dayId = typeof id === 'number' ? String(id) : id;
       deleteScheduleAPI(day, dayId);
       window.alert('Schedule canceled!');
+      changeRefreshState();
     }
   }
